+++ kotlinx-coroutines-core/jvm/test/scheduling/DispatcherParallelismAdjustmentTest.kt
@@ -34,7 +34,7 @@ class DispatcherParallelismAdjustmentTest : SchedulerTestBase() {
             // release), so this extra task can only start if adjustParallelism(1) genuinely grants a
             // new permit.
             assertEquals(
-                1.toByte(),
+                1,
                 dispatcher.tryAdjustParallelism(1),
                 "a single unit of headroom should be granted in full"
             )
@@ -46,7 +46,7 @@ class DispatcherParallelismAdjustmentTest : SchedulerTestBase() {
             )
 
             assertEquals(
-                (-1).toByte(),
+                -1,
                 dispatcher.tryAdjustParallelism(-1),
                 "the previously granted unit of headroom should be reclaimed in full"
             )
@@ -61,8 +61,8 @@ class DispatcherParallelismAdjustmentTest : SchedulerTestBase() {
         maxPoolSize = 64
         // tryAdjustParallelism(delta) increments/decrements one unit at a time internally; make sure it
         // does so exactly `delta` times rather than off by one in either direction.
-        assertEquals(3.toByte(), dispatcher.tryAdjustParallelism(3), "adjustParallelism(3) should grant exactly 3 extra units of parallelism")
-        assertEquals((-3).toByte(), dispatcher.tryAdjustParallelism(-3), "adjustParallelism(-3) should reclaim exactly the 3 units granted above")
+        assertEquals(3, dispatcher.tryAdjustParallelism(3), "adjustParallelism(3) should grant exactly 3 extra units of parallelism")
+        assertEquals(-3, dispatcher.tryAdjustParallelism(-3), "adjustParallelism(-3) should reclaim exactly the 3 units granted above")
     }
 
     @Test
@@ -70,7 +70,7 @@ class DispatcherParallelismAdjustmentTest : SchedulerTestBase() {
         corePoolSize = 2
         // No matching adjustParallelism(1) beforehand -> zero legitimate headroom, so this must not
         // shrink the pool below corePoolSize.
-        assertEquals(0.toByte(), dispatcher.tryAdjustParallelism(-1), "there is no outstanding compensation to reclaim, so nothing should be adjusted")
+        assertEquals(0, dispatcher.tryAdjustParallelism(-1), "there is no outstanding compensation to reclaim, so nothing should be adjusted")
 
         val started = CountDownLatch(corePoolSize)
         val release = CountDownLatch(1)
@@ -105,7 +105,7 @@ class DispatcherParallelismAdjustmentTest : SchedulerTestBase() {
         }
         assertTrue(started.await(10, TimeUnit.SECONDS))
 
-        assertEquals(1.toByte(), soft.tryAdjustParallelism(1), "a single unit of headroom should be granted in full")
+        assertEquals(1, soft.tryAdjustParallelism(1), "a single unit of headroom should be granted in full")
         val extraStarted = CountDownLatch(1)
         soft.dispatch(EmptyCoroutineContext, Runnable { extraStarted.countDown() })
         assertTrue(
@@ -113,7 +113,7 @@ class DispatcherParallelismAdjustmentTest : SchedulerTestBase() {
             "adjustParallelism(1) should let one extra task run concurrently on the soft-limited view"
         )
 
-        assertEquals((-1).toByte(), soft.tryAdjustParallelism(-1), "the previously granted unit of headroom should be reclaimed in full")
+        assertEquals(-1, soft.tryAdjustParallelism(-1), "the previously granted unit of headroom should be reclaimed in full")
         release.countDown()
     }
 
@@ -124,7 +124,7 @@ class DispatcherParallelismAdjustmentTest : SchedulerTestBase() {
 
         // A soft-limited view must never drop below its initial parallelism on its own: there is no
         // outstanding compensation to reclaim, so this must be a no-op.
-        assertEquals(0.toByte(), soft.tryAdjustParallelism(-1), "a soft-limited view must not shrink below its initial parallelism")
+        assertEquals(0, soft.tryAdjustParallelism(-1), "a soft-limited view must not shrink below its initial parallelism")
 
         val started = CountDownLatch(parallelism)
         val release = CountDownLatch(1)
@@ -165,7 +165,7 @@ class DispatcherParallelismAdjustmentTest : SchedulerTestBase() {
                 "sanity check: the second task should still be queued, since no permit was available for it"
             )
 
-            assertEquals(1.toByte(), soft.tryAdjustParallelism(1), "a single unit of headroom should be granted in full")
+            assertEquals(1, soft.tryAdjustParallelism(1), "a single unit of headroom should be granted in full")
             // The sole worker above is still stuck on release.await() (simulating a deadlocked task), so the
             // only way the already-queued task can start now is if adjustParallelism(1) itself kicks the queue.
             assertTrue(
@@ -186,7 +186,7 @@ class DispatcherParallelismAdjustmentTest : SchedulerTestBase() {
         val executor = Executors.newFixedThreadPool(2)
         try {
             repeat(10_000 * stressTestMultiplierSqrt) {
-                assertEquals(1.toByte(), soft.tryAdjustParallelism(1), "should always be able to grant a single unit of headroom")
+                assertEquals(1, soft.tryAdjustParallelism(1), "should always be able to grant a single unit of headroom")
 
                 val barrier = CyclicBarrier(2)
                 val futures = List(2) {
